Integrated Mechanisms with Timeout
We also need to make a template for the AutoPaths.

+++ 2024-Competition-Code/src/main/java/frc/robot/subsystems/Shooter.java
@@ -178,7 +178,7 @@ public class Shooter extends SubsystemBase {
     this.lowerRPM = val;
   }
 
-  public void setIntakePower(int val){
+  public void setIntakePower(double val){
     this.intakePower = val;
   }
